Migrated bot over to a Cog system. Added all functionality into Cogs TODO: Fix Todo/Trello feature which is the only thing not working.
