+++ bot.py
@@ -1,0 +1,63 @@
+import discord
+import os
+import logging
+from discord.ext import commands
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.DEBUG)
+handler = logging.FileHandler(filename='./discordbot.log', encoding='utf-8', mode='w')
+handler.setFormatter(logging.Formatter('%(asctime)s:%(levelname)s:%(name)s: %(message)s'))
+
+
+def get_prefix(bot, message):
+    """A callable Prefix for our bot. This could be edited to allow per server prefixes."""
+
+    # Notice how you can use spaces in prefixes. Try to keep them simple though.
+    prefixes = ['?', '!']
+
+    # Check to see if we are outside of a guild. e.g DM's etc.
+    if not message.guild:
+        # Only allow ? to be used in DMs
+        return '?'
+
+    # If we are in a guild, we allow for the user to mention us or use any of the prefixes in our list.
+    return commands.when_mentioned_or(*prefixes)(bot, message)
+
+
+# Cogs need to be located in cogs folder and added here to load on startup
+initial_extensions = ['cogs.random',
+                      'cogs.reddit',
+                      'cogs.youtube',
+                      'cogs.todo',
+                      'cogs.challenges']
+
+bot = commands.Bot(command_prefix=get_prefix, description='ZTM Python Discord Bot')
+
+# Loading cogs
+if __name__ == '__main__':
+    for extension in initial_extensions:
+        bot.load_extension(extension)
+
+
+@bot.event
+async def on_ready():
+
+    print(f'\n\nLogged in as: {bot.user.name} - {bot.user.id}\nVersion: {discord.__version__}\n')
+
+    # Changes our bots Playing Status. type=1(streaming) for a standard game you could remove type and url.
+    await bot.change_presence(activity=discord.Game(name='The Witcher 3'))
+    print(f'Successfully logged in and booted...!')
+
+
+# member has joined server
+@bot.event
+async def on_member_join(member):
+    print(f'{member} has joined the server')
+
+
+# member has left server
+@bot.event
+async def on_member_remove(member):
+    print(f'{member} has left the server')
+
+bot.run(os.environ['DISCORD_TOKEN'], bot=True, reconnect=True)
